refactor cache into separate class so it follows BearerTokenAuthenticationPolicy + AccessTokenCache pattern
